Make packages and add normal login flow
Login pages are now good and we can move on to authenticating users
