+++ app/src/main/AndroidManifest.xml
@@ -18,18 +18,18 @@
         android:screenOrientation="portrait"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".SplashActivity">
+        <activity android:name=".activity.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".HomeActivity" />
-        <activity android:name=".ScheduleActivity" />
-        <activity android:name=".NotificationsActivity" />
-        <activity android:name=".GitHubLoginActivity" />
-        <activity android:name=".LoginChooserActivity" />
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".activity.HomeActivity" />
+        <activity android:name=".activity.ScheduleActivity" />
+        <activity android:name=".activity.NotificationsActivity" />
+        <activity android:name=".activity.login.GitHubLoginActivity" />
+        <activity android:name=".activity.login.LoginChooserActivity" />
+        <activity android:name=".activity.login.LoginActivity" />
     </application>
 
 </manifest>
