--- app/src/main/java/org/hackillinois/android/NotificationsActivity.java
@@ -1,13 +1,0 @@
-package org.hackillinois.android;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class NotificationsActivity extends DrawerActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_notifications);
-    }
-}
