--- app/src/main/java/org/hackillinois/android/ScheduleActivity.java
@@ -1,13 +1,0 @@
-package org.hackillinois.android;
-
-import android.support.v7.app.AppCompatActivity;
-import android.os.Bundle;
-
-public class ScheduleActivity extends DrawerActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_schedule);
-    }
-}
