+++ app/src/main/java/org/hackillinois/android/activity/HackillinoisActivity.java
@@ -1,4 +1,4 @@
-package org.hackillinois.android;
+package org.hackillinois.android.activity;
 
 import android.content.Context;
 import android.support.v7.app.AppCompatActivity;
