+++ app/src/main/java/org/hackillinois/android/activity/ScheduleActivity.java
@@ -1,0 +1,14 @@
+package org.hackillinois.android.activity;
+
+import android.os.Bundle;
+
+import org.hackillinois.android.R;
+
+public class ScheduleActivity extends DrawerActivity {
+
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		setContentView(R.layout.activity_schedule);
+	}
+}
