+++ app/src/main/java/org/hackillinois/android/activity/login/GitHubLoginActivity.java
@@ -1,9 +1,13 @@
-package org.hackillinois.android;
+package org.hackillinois.android.activity.login;
 
 import android.os.Bundle;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.Toast;
+
+import org.hackillinois.android.R;
+import org.hackillinois.android.activity.HackillinoisActivity;
+import org.hackillinois.android.api.ApiEndpoints;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -22,7 +26,8 @@ public class GitHubLoginActivity extends HackillinoisActivity {
 		githubWebview.setWebViewClient(new WebViewClient() {
 			@Override
 			public boolean shouldOverrideUrlLoading(WebView view, String url) {
-				if(url.startsWith("https://hackillinois.org/auth/?code=") || url.contains("?code=")) {
+				if (url.startsWith("https://hackillinois.org/auth/?code=") || url.contains("?code=")) {
+					// todo clean up url checking
 					//we got the code
 					Toast.makeText(getApplicationContext(), "Got " + url, Toast.LENGTH_LONG).show();
 					return true;
