+++ app/src/main/java/org/hackillinois/android/activity/login/LoginActivity.java
@@ -1,0 +1,15 @@
+package org.hackillinois.android.activity.login;
+
+import android.os.Bundle;
+
+import org.hackillinois.android.R;
+import org.hackillinois.android.activity.HackillinoisActivity;
+
+public class LoginActivity extends HackillinoisActivity {
+
+	@Override
+	protected void onCreate(Bundle savedInstanceState) {
+		super.onCreate(savedInstanceState);
+		setContentView(R.layout.activity_login);
+	}
+}
