+++ app/src/main/res/layout/activity_drawer.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context="org.hackillinois.android.DrawerActivity">
+    tools:context="org.hackillinois.android.activity.DrawerActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
