+++ app/src/main/res/layout/activity_github_login.xml
@@ -6,4 +6,4 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="org.hackillinois.android.GitHubLoginActivity"/>
+    tools:context="org.hackillinois.android.activity.login.GitHubLoginActivity"/>
