+++ app/src/main/res/layout/activity_schedule.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="org.hackillinois.android.ScheduleActivity">
+    tools:context="org.hackillinois.android.activity.ScheduleActivity">
 
 </android.support.constraint.ConstraintLayout>
